requirements.txt
absl-py==0.9.0
appdirs==1.4.4
astor==0.8.1
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cmake==3.17.3
distlib==0.3.0
dlib @ https://pypi.python.org/packages/da/06/bd3e241c4eb0a662914b3b4875fc52dd176a9db0d4a2c915ac2ad8800e9e/dlib-19.7.0-cp36-cp36m-win_amd64.whl
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.0.12
Flask==1.1.2
gast==0.3.3
google-api-core==1.17.0
google-auth==1.16.0
google-cloud-core==1.3.0
google-cloud-speech==1.3.2
google-cloud-storage==1.28.1
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.29.0
h5py==2.10.0
idna==2.9
importlib-metadata==1.6.0
importlib-resources==1.5.0
imutils==0.5.3
itsdangerous==1.1.0
Jinja2==2.11.2
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
MarkupSafe==1.1.1
numpy==1.18.5
opencv-python==4.2.0.34
Pillow==7.1.2
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
rsa==4.0
scipy==1.4.1
six==1.15.0
tensorboard==1.12.2
tensorflow-gpu==1.12.0
termcolor==1.1.0
urllib3==1.25.9
virtualenv==20.0.21
virtualenvwrapper-win==1.2.6
Werkzeug==1.0.1
zipp==3.1.0
